Replace image download URLs based on API host
This script is by default only available on the server container,
however inside container because of hairpin problem we cannot use
image URLs as set in the pillar. This PR modifies download URL to
use API host instead (and disable SSL verification because of it).

Also removes socket.getFqdn for delta url and use host from image
url instead.